finished exercise 2 chapter 7

+++ Controllers/DogsController.cs
@@ -1,8 +1,10 @@
 ﻿using DogGo.Models;
 using DogGo.Repositories;
+using Humanizer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace DogGo.Controllers
 {
@@ -12,6 +14,12 @@ namespace DogGo.Controllers
 
         private readonly IDogRepository _dogRepo;
 
+        private int GetCurrentUserId()
+        {
+            string id = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            return int.Parse(id);
+        }
+
         // Constructor / ASP.NET will give us an instance of our Dog Repository when creating an instance of DogsController. This is called "Dependency Injection"
         public DogsController(IDogRepository dogRepository)
         {
@@ -20,11 +28,21 @@ namespace DogGo.Controllers
 
 
 
-        // GET: DogsController
+        //// GET: DogsController Chapter 3
+        //public ActionResult Index()
+        //{
+        //    List<Dog> dogslist = _dogRepo.GetAllDogs();
+        //    return View(dogslist);
+        //}
+
+        [Authorize]//stops unauthenticated users from accessing /dogs...must be logged in 
         public ActionResult Index()
         {
-            List<Dog> dogslist = _dogRepo.GetAllDogs();
-            return View(dogslist);
+            int ownerId = GetCurrentUserId();
+
+            List<Dog> dogs = _dogRepo.GetDogsByOwnerId(ownerId);
+
+            return View(dogs);
         }
 
         // GET: DogsController/Details/5
@@ -35,6 +53,7 @@ namespace DogGo.Controllers
 
         // GET: DogsController/Create
         [HttpGet]//Specified the HTTP attributes ([HttpGet] and [HttpPost])=> make sure correct method invoked based on the request type since have 2 method with name create.
+        [Authorize] //Stops unauthenticated users from accessing /dogs/create route...will be redirected to login page
         public ActionResult Create()
         {
             return View();
@@ -63,13 +82,15 @@ namespace DogGo.Controllers
         //}
 
 
-        // POST: DogController/Create
+        // POST: DogController/Create:Defaulting the OwnerId when creating dogs =>so user dont have to add ownerID when creating a new dog profile
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create(Dog dog)
         {
             try
             {
+                // update the dogs OwnerId to the current user's Id
+                dog.OwnerId = GetCurrentUserId();
                 _dogRepo.AddDog(dog);
 
                 return RedirectToAction("Index");
@@ -82,26 +103,60 @@ namespace DogGo.Controllers
 
         // GET: DogsController/Edit/5
 
+        [Authorize]
         public ActionResult Edit(int id)
         {
+
             Dog dog = _dogRepo.GetDogById(id);
+            
 
             if (dog == null)
             {
-                return NotFound();
+                return NotFound("Ooops! Dog Not Found Database");
             }
+            // Check if the current user owns the dog
+            int currentUserId = GetCurrentUserId();
+            
+            if (dog.OwnerId != currentUserId)
+            {
+                return NotFound("Not Authorized!!!");
+            }
+
             return View(dog);
         }
-
+          
         // POST: DogsController/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult Edit(int id, Dog dog)
         {
             try
             {
-                _dogRepo.UpdateDog(dog);
-                return RedirectToAction(nameof(Index));
+                Dog dogInDB = _dogRepo.GetDogById(id);
+
+                if (dogInDB == null)
+                {
+                    return NotFound("OOPS! Dog not fough in our database!");
+                }
+                // Check if the current user owns the dog
+                int currentUserId = GetCurrentUserId();
+
+                if (dogInDB.OwnerId != currentUserId)
+                {
+                    return NotFound("Not Authorized!!!");
+                }
+                else
+                {
+                    dogInDB.Name = dog.Name;
+                    dogInDB.Breed = dog.Breed;
+                    dogInDB.ImageUrl = dog.ImageUrl;
+                    dogInDB.Notes = dog.Notes;
+                    dogInDB.OwnerId = currentUserId;
+
+                    _dogRepo.UpdateDog(dogInDB);
+                    return RedirectToAction(nameof(Index));
+                }
             }
             catch
             {
@@ -111,20 +166,53 @@ namespace DogGo.Controllers
 
 
         // GET: DogsController/Delete/5
-      
+
+        [Authorize]
         public ActionResult Delete(int id)
         {
             Dog dog = _dogRepo.GetDogById(id);
+
+
+            if (dog == null)
+            {
+                return NotFound("Oops! Dog not found!");
+            }
+            // Check if the current user owns the dog
+            int currentUserId = GetCurrentUserId();
+
+            if (dog.OwnerId != currentUserId)
+            {
+                return NotFound("Accesss Denied!");
+            }
+
             return View(dog);
         }
+
 
         // POST: DogsController/Delete/5
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult Delete(int id, Dog dog)
         {
+            Dog dogInDB = _dogRepo.GetDogById(id);
+
+
+            if (dogInDB == null)
+            {
+                return NotFound();
+            }
+            // Check if the current user owns the dog
+            int currentUserId = GetCurrentUserId();
+
+            if (dogInDB.OwnerId != currentUserId)
+            {
+                return NotFound();
+            }
+
             try
             {
+
                 _dogRepo.DeleteDog(id);
                 return RedirectToAction("Index");
             }
@@ -135,3 +223,4 @@ namespace DogGo.Controllers
         }
     }
 }
+
